backend/requirements.txt
# Autosell.mx Backend Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Google APIs
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
gspread>=5.12.0

# Social Media APIs
facebook-business>=18.0.0
tweepy>=4.14.0

# HTTP Client
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# Data Processing
pandas>=2.1.3
numpy>=1.25.2
pillow>=10.1.0

# Caching & Performance
redis>=5.0.1
aioredis>=2.0.1

# Authentication & Security
bcrypt>=4.1.2
cryptography>=41.0.7

# File Handling
aiofiles>=23.2.1

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Development Tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
python-dateutil>=2.8.2
pytz>=2023.3

# Image Processing
opencv-python>=4.8.1.78
imageio>=2.33.1

# Business Intelligence
matplotlib>=3.8.2
seaborn>=0.13.0
scikit-learn>=1.3.2

# Web Scraping (for market analysis)
beautifulsoup4>=4.12.2
selenium>=4.15.2

# Background Tasks
celery>=5.3.4

# Email & Notifications
fastapi-mail>=1.4.1
jinja2>=3.1.2

# SMS (Twilio)
twilio>=8.10.0

# Async Support
aio-pika>=9.3.1

# Health Checks
healthcheck>=1.3.0

# Rate Limiting
slowapi>=0.1.9

# CORS
fastapi-cors>=0.0.6
